move the gitfs_log fixture to integrations/base for reuse
