+++ tests/integrations/current/test_write.py
@@ -14,7 +14,6 @@
 
 
 import collections
-import contextlib
 import os
 import pytest
 import re
@@ -22,69 +21,7 @@ import shutil
 import string
 import time
 
-from tests.integrations.base import BaseTest, pull
-
-
-class GitFSLog(object):
-    def __init__(self, file_descriptor):
-        self._partial_line = None
-        self.line_buffer = collections.deque()
-        self.file_descriptor = file_descriptor
-
-    def get_line(self):
-        if not self.line_buffer:
-            self._read_data()
-        if self.line_buffer:
-            return self.line_buffer.popleft()
-
-    def _read_data(self):
-        data = os.read(self.file_descriptor, 2048).splitlines(True)
-        if not data:
-            return
-        if self._partial_line:
-            data[0] = self._partial_line + data[0]
-        if not data[-1].endswith("\n"):
-            self._partial_line = data[-1]
-            data = data[:-1]  # discard the partial line
-        else:
-            self._partial_line = None
-        self.line_buffer.extend(data)
-
-    def clear(self):
-        while os.read(self.file_descriptor, 2048):
-            pass
-        self._partial_line = None
-        self.line_buffer = collections.deque()
-
-    def __call__(self, expected):
-        @contextlib.contextmanager
-        def log_context(gitfs_log):
-            gitfs_log.clear()
-            yield
-            gitfs_log.expect(expected)
-        return log_context(self)
-
-    def expect(self, expected, timeout=5, pollfreq=0.1):
-        expected = re.compile(".*"+expected)
-        return self.expect_re(expected, timeout, pollfreq)
-
-    def expect_re(self, expected, timeout=5, pollfreq=0.2):
-        elapsed = 0
-        while elapsed < timeout:
-            line = self.get_line()
-            while line is None:
-                time.sleep(pollfreq)
-                elapsed += pollfreq
-                line = self.get_line()
-            if expected.match(line):
-                return line
-        raise AssertionError("expected '{}' not found in stream".format(expected))
-
-
-@pytest.fixture(scope='session')
-def gitfs_log():
-    return GitFSLog(os.open("log.txt", 0))
-
+from tests.integrations.base import BaseTest, pull, gitfs_log
 
 
 class TestWriteCurrentView(BaseTest):
